Apply sandbox badge
This PR applies sandbox badge to the repository to signal
the level of maturity of the project as agreed by Eiffel
Bootstrap Technical Committee.

See https://github.com/eiffel-community/community/blob/master/PROJECT_LIFECYCLE.md#project-lifecycle-stages
for the details of project lifecycle.

See https://hackmd.io/SCImga0nS1qSh3QvsEOAVQ?view for the
discussion that took place during past Eiffel Bootstrap
Technical Committee Meetings.

https://github.com/eiffel-community/community/issues/85

Include sandbox badge in readme file to increase the visibility.

N/A

Having information about the status of projects upfront makes it easier for potential users to estimate the effort needed for deployment.

If we're not careful it might give the wrong impression and make users hesitant to try out, and contribute to, the development of new tools.

<!-- Sign the below certificate of origin, using your full name and e-mail address. -->
<!-- The certificate is copied from https://developercertificate.org/ -->

Developer's Certificate of Origin 1.1

By making a contribution to this project, I certify that:

(a) The contribution was created in whole or in part by me and I
    have the right to submit it under the open source license
    indicated in the file; or

(b) The contribution is based upon previous work that, to the best
    of my knowledge, is covered under an appropriate open source
    license and I have the right under that license to submit that
    work with modifications, whether created in whole or in part
    by me, under the same open source license (unless I am
    permitted to submit under a different license), as indicated
    in the file; or

(c) The contribution was provided directly to me by some other
    person who certified (a), (b) or (c) and I have not modified
    it.

(d) I understand and agree that this project and the contribution
    are public and that a record of the contribution (including all
    personal information I submit with it, including my sign-off) is
    maintained indefinitely and may be redistributed consistent with
    this project or the open source license(s) involved.

Signed-off-by: Fatih Degirmenci fatih.degirmenci@est.tech